Ignore package-lock.json, try using rn/yarn_install on circleci, move RC init to app.tsx
