+++ .gitignore
@@ -9,6 +9,7 @@ node_modules/
 npm-debug.log
 yarn-error.log
 **/yarn.lock
+**/package-lock.json
 
 # Xcode
 #
